+++ AppCutAudio/AppCutAudio/MainWindow.xaml.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.Win32;
 using NAudio.Wave;
-using OxyPlot.Series;
-using OxyPlot;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,6 +22,8 @@ using System.Collections.ObjectModel;
 using AppCutAudio.Models;
 using System.Windows.Threading;
 using System.ComponentModel;
+using AppCutAudio.Vistas;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.Window;
 
 namespace AppCutAudio
 {
@@ -40,6 +40,7 @@ namespace AppCutAudio
         private readonly DispatcherTimer _timer = new DispatcherTimer();
         private int _currentImageIndex = 0;
         private string[] _imagePaths = { "/Images/Home.jpeg", "/Images/Home1.jpeg", "/Images/Home2.jpeg", "/Images/Home3.jpeg" };
+        private string ProyectName;
         public ObservableCollection<AudioFile> ListaDeAudios { get; set; }
         public MainWindow()
         {
@@ -61,29 +62,48 @@ namespace AppCutAudio
 
         private async void btnAbrir_Click(object sender, RoutedEventArgs e)
         {
-            string url = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "espectrograma.png");
-            File.Delete(url);
-            btnAbrir.IsEnabled = false;
-            btnLimpiar.IsEnabled = true;
-            Microsoft.Win32.OpenFileDialog openFileDialog = new Microsoft.Win32.OpenFileDialog();
-            openFileDialog.Filter = "Audio files (*.mp3;*.m4a;*.wav)|*.mp3;*.m4a;*.wav|All files (*.*)|*.*";
-            if (openFileDialog.ShowDialog() == true)
+            var dialogo = new Vistas.Dialog();
+            if (dialogo.ShowDialog() == true)
             {
-                using (Mp3FileReader reader = new Mp3FileReader(openFileDialog.FileName))
+                if (dialogo.NombreDelProyecto.ToLower() == "demo")
                 {
-                    TimeSpan duration = reader.TotalTime;
-                    if (duration.TotalMinutes > 3)
+                    MessageBox.Show("El nombre Demo esta reservado por el programa, prueba con otro nombre.");
+                    btnAbrir_Click(sender, e);
+                }
+                else
+                {
+                    ProyectName = dialogo.NombreDelProyecto;
+                    string carpetaProyecto = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, ProyectName);
+                    if (Directory.Exists(carpetaProyecto))
+                        Directory.Delete(carpetaProyecto, true);
+
+                    string url = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "espectrograma.png");
+                    File.Delete(url);
+
+                    btnAbrir.IsEnabled = false;
+                    btnLimpiar.IsEnabled = true;
+                    Microsoft.Win32.OpenFileDialog openFileDialog = new Microsoft.Win32.OpenFileDialog();
+                    openFileDialog.Filter = "Audio files (*.mp3;*.m4a;*.wav)|*.mp3;*.m4a;*.wav|All files (*.*)|*.*";
+                    if (openFileDialog.ShowDialog() == true)
                     {
-                        btnProcesar.IsEnabled = true;
-                        overlay.Visibility = Visibility.Visible;
-                        pathAudio = openFileDialog.FileName;
-                        fileName = openFileDialog.SafeFileName;
-                        await ShowAudio(openFileDialog);
-                        InputsHideShow(Visibility.Visible);
-                    }
-                    else
-                    {
-                        MessageBox.Show("El archivo de audio seleccionado tiene menos de 3 minutos de duración.");
+                        Mp3FileReader mp3FileReader = new Mp3FileReader(openFileDialog.FileName);
+                        using (Mp3FileReader reader = mp3FileReader)
+                        {
+                            TimeSpan duration = reader.TotalTime;
+                            if (duration.TotalMinutes > 1)
+                            {
+                                btnProcesar.IsEnabled = true;
+                                overlay.Visibility = Visibility.Visible;
+                                pathAudio = openFileDialog.FileName;
+                                fileName = openFileDialog.SafeFileName;
+                                await ShowAudio(openFileDialog);
+                                InputsHideShow(Visibility.Visible);
+                            }
+                            else
+                            {
+                                MessageBox.Show("El archivo de audio seleccionado tiene menos de 1 minutos de duración.");
+                            }
+                        }
                     }
                 }
             }
@@ -99,10 +119,7 @@ namespace AppCutAudio
                 Genre = txtGenero.Text,
                 Year = txtYear.Text
             };
-            string pathSegmentos = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, fileName);
-            if(Directory.Exists(pathSegmentos))
-                Directory.Delete(pathSegmentos, true);
-            Directory.CreateDirectory(pathSegmentos);
+            string pathSegmentos = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, ProyectName);
 
             // await ProcesarAudio(); //python
             AudioDetector audioDetector = new AudioDetector();
@@ -128,15 +145,21 @@ namespace AppCutAudio
             PlayButton.Visibility = Visibility.Collapsed;
             StopButton.Visibility = Visibility.Collapsed;
             btnProcesar.IsEnabled = false;
-            string pathSegmentos = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, fileName);
-            if( Directory.Exists(pathSegmentos))
-                Directory.Delete(pathSegmentos, true);
+            if (!string.IsNullOrEmpty(fileName))
+            {
+                string pathSegmentos = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, fileName);
+                if (Directory.Exists(pathSegmentos))
+                    Directory.Delete(pathSegmentos, true);
+            }
             InputsHideShow(Visibility.Collapsed);
+            lstAudios.ItemsSource = null;
             ListaDeAudios = new ObservableCollection<AudioFile>();
+            lstAudios.ItemsSource = ListaDeAudios;
         }
 
         private void PlayButton_Click(object sender, RoutedEventArgs e)
         {
+            waveOut?.Stop();
             PlayFromUrl(pathAudio);
         }
 
@@ -147,7 +170,18 @@ namespace AppCutAudio
 
         private void EditItemListButton_Click(object sender, RoutedEventArgs e)
         {
-            
+            // Obtén el elemento seleccionado
+            Button playButton = (Button)sender;
+            var audio = (AudioFile)playButton.Tag;
+
+            // Encuentra el índice del audio en la lista
+            int index = ListaDeAudios.IndexOf(audio);
+
+            TrackDetails trackDetailsWindow = new TrackDetails(audio);
+            if (trackDetailsWindow.ShowDialog() == true)
+            {
+                ListaDeAudios[index] = trackDetailsWindow.AudioUpdate;
+            }
         }
 
         private void PlayItemListButton_Click(object sender, RoutedEventArgs e)
@@ -155,61 +189,81 @@ namespace AppCutAudio
             // Obtén el elemento seleccionado
             Button playButton = (Button)sender;
             var audio = (AudioFile)playButton.Tag;
+            waveOut?.Stop();
             PlayFromUrl(audio.RutaArchivo);
         }
         private void StopItemListButton_Click(object sender, RoutedEventArgs e)
         {
             waveOut?.Stop();
         }
+        private void BtnExportar_Click(object sender, RoutedEventArgs e)
+        {
+            // Mostrar un MessageBox al usuario
+            MessageBox.Show("Por favor, elige la carpeta donde guardar tus archivos generados.");
+
+            // Crear una nueva instancia de FolderBrowserDialog
+            System.Windows.Forms.FolderBrowserDialog folderBrowserDialog = new System.Windows.Forms.FolderBrowserDialog();
+
+            // Mostrar el FolderBrowserDialog y obtener el resultado
+            System.Windows.Forms.DialogResult result = folderBrowserDialog.ShowDialog();
+
+            // Si el usuario seleccionó una carpeta y presionó el botón OK
+            if (result == System.Windows.Forms.DialogResult.OK)
+            {
+                // Obtener la carpeta seleccionada
+                string selectedPath = folderBrowserDialog.SelectedPath;
+
+                foreach (AudioFile item in ListaDeAudios)
+                {
+                    AudioHelper.SaveMp3WithID3Tag(item.RutaArchivo, $"{selectedPath}/{item.NombreArchivo}", item.GetTagData());
+                }
+            }
+        }
+
+        private async void btnDemo_Click(object sender, RoutedEventArgs e)
+        {
+            PlayButton.Visibility = Visibility.Visible;
+            StopButton.Visibility = Visibility.Visible;
+            btnAbrir.IsEnabled = false;
+            btnLimpiar.IsEnabled = true;
+            btnExportar.IsEnabled = false;
+
+            string pathDemo = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Demo");
+            string finalimage = System.IO.Path.Combine(pathDemo, "finalResult.png");
+            pathAudio = $"{pathDemo}/demo.mp3";
+
+            BitmapImage finalBitmap = new BitmapImage(new Uri(finalimage));
+            audioImage.Source = finalBitmap;
+
+            await CargarListaDemoAsync(pathDemo);
+        }
 
         #region Helper
         private async Task ShowAudio(Microsoft.Win32.OpenFileDialog openFileDialog)
         {
-            string pythonScriptPath = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "PythonScripts", "AudioGrafic.py");
+            string carpetaProyecto = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, ProyectName);
             string filePath = openFileDialog.FileName;
-
-            if (!File.Exists(pythonScriptPath))
-            {
-                MessageBox.Show("El archivo de script de Python no se encontró.");
-                return;
-            }
 
             try
             {
-                using (Process process = new Process())
+                //AudioHelper.DividirAudio(filePath, carpetaProyecto, new TimeSpan(0, 30, 0));
+                AudioHelper.CrearDemo(filePath, carpetaProyecto, new TimeSpan(0, 3, 0));
+
+                string[] partes = Directory.GetFiles(carpetaProyecto);
+
+                if (partes.Any())
                 {
-                    process.StartInfo.FileName = "python";
-                    process.StartInfo.Arguments = $"\"{pythonScriptPath}\" \"{filePath}\"";
-                    process.StartInfo.RedirectStandardOutput = true;
-                    process.StartInfo.UseShellExecute = false;
-                    process.StartInfo.CreateNoWindow = true;
+                    string finalimage = System.IO.Path.Combine(carpetaProyecto, "finalResult.png");
 
-                    process.Start();
-                    string output = await process.StandardOutput.ReadToEndAsync();
+                    AudioDetector detector = new AudioDetector();
+                    await Task.Run(() =>
+                    {
+                        detector.GenerarEspectrograma(partes[0], finalimage);
+                    });
 
-                    // Verificar si hay errores de salida
-                    if (!string.IsNullOrWhiteSpace(output))
-                    {
-                        MessageBox.Show($"El script de Python produjo la siguiente salida: {output}");
-                    }
-
-                    // Haz algo con la salida del script de Python (si es necesario)
-                    // Carga la imagen y asígnala al control Image
-                    string imagePath = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "espectrograma.png");
-                    if (File.Exists(imagePath))
-                    {
-                        BitmapImage bitmap = new BitmapImage(new Uri(imagePath));
-                        audioImage.Source = bitmap;
-                    }
-                    else
-                    {
-                        MessageBox.Show("La imagen generada por el script de Python no se encontró.");
-                    }
+                    BitmapImage finalBitmap = new BitmapImage(new Uri(finalimage));
+                    audioImage.Source = finalBitmap;
                 }
-            }
-            catch (Win32Exception ex)
-            {
-                MessageBox.Show($"Error al iniciar el proceso de Python: {ex.Message}");
             }
             catch (Exception ex)
             {
@@ -220,46 +274,6 @@ namespace AppCutAudio
             PlayButton.Visibility = Visibility.Visible;
             StopButton.Visibility = Visibility.Visible;
         }
-
-
-        private async Task ProcesarAudio()
-        {
-            // Obtén la ruta completa al script de Python
-            string pythonScriptPath = System.IO.Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "PythonScripts", "AudioCut.py");
-
-            // Invoca el script de Python con la ruta del archivo como argumento
-            ProcessStartInfo psi = new ProcessStartInfo
-            {
-                FileName = "python", // Asegúrate de que "python" esté en tu PATH
-                Arguments = $"\"{pythonScriptPath}\" \"{pathAudio}\"", // Pasa la ruta del archivo como argumento
-                RedirectStandardOutput = true,
-                UseShellExecute = false,
-                CreateNoWindow = true
-            };
-
-            using (Process process = new Process { StartInfo = psi })
-            {
-                try
-                {
-                    process.Start();
-                    string output = await process.StandardOutput.ReadToEndAsync();
-
-                    int posicionError = output.IndexOf("error", StringComparison.OrdinalIgnoreCase);
-
-                    if (posicionError >= 0)
-                    {
-                        string errorPart = output.Substring(posicionError);
-                        errorPart = errorPart.Trim();
-                        throw new Exception($"Error: {errorPart}");
-                    }
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message);
-                }
-            }
-        }
-
         private void InputsHideShow(Visibility visibility)
         {
             txtAlbum.Visibility = visibility;
@@ -291,7 +305,7 @@ namespace AppCutAudio
         }
         private void SetBackgroundImage(string imagePath)
         {
-            imagePath= $"{AppDomain.CurrentDomain.BaseDirectory}{imagePath}";
+            imagePath = $"{AppDomain.CurrentDomain.BaseDirectory}{imagePath}";
             try
             {
                 imgBrush.ImageSource = new BitmapImage(new Uri(imagePath, UriKind.Relative));
@@ -300,6 +314,30 @@ namespace AppCutAudio
             {
                 // Manejar errores al cargar la imagen
                 MessageBox.Show($"Error al cargar la imagen: {ex.Message}");
+            }
+        }
+
+        private async Task CargarListaDemoAsync(string pathSegmentos)
+        {
+            tagData = new NAudio.Lame.ID3TagData()
+            {
+                Album = "CHUBBY MOSH Trilogía",
+                Artist = "David Fau Casquel y Javier Ribes",
+                Comment = "3º disco de CHUBBY MOSH",
+                Genre = "rock, metal, punk, fussion",
+                Year = "2018"
+            };
+
+            AudioDetector audioDetector = new AudioDetector();
+
+            ObservableCollection<AudioFile> lista = await audioDetector.ConvertirLista(pathSegmentos, tagData);
+            foreach (var audio in lista)
+            {
+                audio.Title = audio.NombreArchivo.Remove(0, 3).Replace(".mp3", "");
+                Application.Current.Dispatcher.Invoke(() =>
+                {
+                    ListaDeAudios.Add(audio);
+                });
             }
         }
         #endregion
